+++ packages/backend-function/src/lambda-shims/resolve_ssm_params.test.ts
@@ -1,5 +1,5 @@
-import { SSM } from '@aws-sdk/client-ssm';
-import { after, describe, it, mock } from 'node:test';
+import { GetParametersCommandInput, SSM } from '@aws-sdk/client-ssm';
+import { afterEach, describe, it, mock } from 'node:test';
 import { internalAmplifyFunctionResolveSsmParams } from './resolve_ssm_params.js';
 import assert from 'node:assert';
 
@@ -8,7 +8,7 @@ void describe('internalAmplifyFunctionResolveSsmParams', () => {
   const client = new SSM();
 
   // reset process.env after test suite to ensure there are no side effects
-  after(() => {
+  afterEach(() => {
     process.env = originalEnv;
   });
 
@@ -42,5 +42,43 @@ void describe('internalAmplifyFunctionResolveSsmParams', () => {
     await internalAmplifyFunctionResolveSsmParams(client);
     assert.equal(mockGetParameters.mock.callCount(), 1);
     assert.equal(process.env[envName], secretValue);
+  });
+
+  void it('paginates when there are more than 10 secrets', async () => {
+    const envName = 'TEST_SECRET';
+    const secretPath = '/test/path';
+    let ssmPaths = {};
+    for (let i = 0; i < 100; i++) {
+      ssmPaths = Object.assign(ssmPaths, {
+        [secretPath + i]: {
+          name: envName + i,
+          sharedSecretPath: '/test/shared/path',
+        },
+      });
+    }
+    process.env.AMPLIFY_SSM_ENV_CONFIG = JSON.stringify(ssmPaths);
+
+    // Let's return the value same as args
+    const mockGetParameters = mock.method(
+      client,
+      'getParameters',
+      (args: GetParametersCommandInput) => {
+        const parameters: unknown[] = [];
+        args.Names?.forEach((name) => {
+          parameters.push({
+            Name: name,
+            Value: name,
+          });
+        });
+        return Promise.resolve({
+          Parameters: parameters,
+        });
+      }
+    );
+    await internalAmplifyFunctionResolveSsmParams(client);
+    assert.equal(mockGetParameters.mock.callCount(), 10);
+    for (let i = 0; i < 100; i++) {
+      assert.equal(process.env[envName + i], secretPath + i);
+    }
   });
 });