+++ packages/backend-deployer/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @aws-amplify/backend-deployer
 
+## 1.1.17
+
+### Patch Changes
+
+- bc07307: Update code with Eslint@8 compliant
+- Updated dependencies [bc07307]
+  - @aws-amplify/platform-core@1.6.2
+
+## 1.1.16
+
+### Patch Changes
+
+- a7ae994: wrap more instances of ENOENT type errors
+- ec7826d: Handle connect ENOMEM error
+- Updated dependencies [ec7826d]
+  - @aws-amplify/platform-core@1.6.1
+
 ## 1.1.15
 
 ### Patch Changes